function/requirements.txt
azure-functions==1.20.0
psycopg2-binary==2.9.9
python-http-client==3.3.7
sendgrid==6.10.0
starkbank-ecdsa==2.2.0